// src/eta/enums/aecparameters.enum.ts
/* eslint-disable @typescript-eslint/camelcase */
// AEC (Scheitholzkessel)
export enum AECParameters {
	Kessel = 8,
	Abgas = 15,
	Boiler = 13,
	Brauchwasser = 164,
	BWT_Ruecklauf = 106,
	BWT_Mitte = 107,
	Pufferladezst = 75,
	Puffer_oben = 12,
	Puffer_mitte = 11,
	Puffer_unten = 10,
	Kesselruecklauf = 9,
	Brenner = 117,
	Kollektor = 62,
	Boiler_oben_Sol = 161,
	Boiler_unten_Sol = 63,
	Aussentemp = 70,
	Vorlauf_MK_1 = 68,
	Raum_MK_1 = 66,
	Vorlauf_MK_2 = 69,
	Raum_MK_2 = 67,
	Vorlauf_MK_3 = 82,
	Raum_MK_3 = 81,
	Vorlauf_MK_4 = 94,
	Raum_MK_4 = 93,
	None = 0,
}