// src/eta/enums/aelkparameters.enum.ts
/* eslint-disable @typescript-eslint/camelcase */
//AELK (Hack, oder Pellets ab 35kW)
export enum AELKParameters {
	Kessel = 20,
	Kesselruecklauf = 112,
	Fremdwaerme = 88,
	Abgas = 19,
	Puffer_oben = 77,
	Puffer_unten = 78,
	Boiler = 21,
	Boiler_unten = 167,
	Aussen = 22,
	Vorlauf_MK_0 = 23,
	Raum_MK_0 = 25,
	Vorlauf_MK_1 = 24,
	Raum_MK_1 = 26,
	Pelletsvorrat = 143,
}