// src/eta/enums/aelparameters.enum.ts
/* eslint-disable @typescript-eslint/camelcase */
//AEL (Pellets bis 30kW)
export enum AELParameters {
	Kessel = 20,
	Kesselvorlauf = 23,
	Fremdwaerme = 88,
	Abgas = 19,
	Puffer_oben = 77,
	Puffer_unten = 78,
	Boiler = 21,
	Boiler_unten = 134,
	Aussen = 22,
	Vorlauf_DK = 23,
	Raum_DK = 25,
	Vorlauf_MK_0 = 24,
	Raum_MK_0 = 124,
	ThermostatHeiss = 96,
	ThermostatKalt = 97,
	Pelletsvorrat = 86,
}